+++ material/src/index.js
@@ -1,3 +1,3 @@
-export { RunlyProvider } from "@runly/core";
+export { RunlyProvider } from "@runly/ui";
 
 export { default as RunProgress } from "./progress";