+++ material/src/progress.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useRunConnection, ProgressText } from "@runly/core";
+import { useRunConnection, ProgressText } from "@runly/ui";
 
 import { Typography, LinearProgress } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";