+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 Runly.js is distributed as a [monorepo][monorepo] using [lerna][lerna]. This means that our `CHANGELOG.md` files are distributed alongside the package itself:
 
-* [`@runly/core`](./core/CHANGELOG.md)
-* [`@runly/react-bootstrap`](./react-bootstrap/CHANGELOG.md)
-* [`@runly/react-material`](./react-material/CHANGELOG.md)
+* [`@runly/ui`](./core/CHANGELOG.md)
+* [`@runly/bootstrap`](./bootstrap/CHANGELOG.md)
+* [`@runly/material`](./material/CHANGELOG.md)
 * [HTML Integration](./html/CHANGELOG.md)
 
 [monorepo]: https://en.wikipedia.org/wiki/Monorepo